set up of home-reoutes and schedule-routes, not fully tested with current main, also api routes removed to not confict with work done by grace
