+++ controllers/api/index.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
 
 const userRoutes = require('./user-routes');
+const exerciseRoutes = require('./exercise-routes');
 
 router.use('/users', userRoutes);
+router.use('/exercise', exerciseRoutes);
 
 module.exports = router;
